Add moomble.py
This is a program which allows a user to send a text message to anyone anywhere in the world in any language you choose (as long as its a language supported by Python's googletrans module!). Its main aim is to bring down language barriers in cross-lingual communication hence the name "moomble", which is a deliberate misspell of "mumble" to indicate that it does the opposite of mumbling i.e. it makes communication more coherent and intelligible.