Add Boys Hate Girls.py
This is a 2D game created using pygame in which you play as a "boy" who has to avoid "girls" falling from the top of the screen. As long as you avoid the girls, your score keeps increasing. But the moment you collide with anyone of them, it's game over for you! It's challenging as the falling girls are all of different sizes and fall from random positions along the top of the screen. I've included a couple of cheats as well (pressing Z reverses the girls' course while pressing X slows them down!) however this temporary respite comes at a cost as using the cheat drops your score to zero!